fix bugs

+++ client/models/usersProducts.js
@@ -7,25 +7,25 @@ const usersProductsDB = {
             .query(sql, [email])
             .then((dbRes) => dbRes.rows);
     },
-    addNewProduct(name, description, address, availability, category, user_id) {
+    addNewProduct(name, description, address, availability, category, userId) {
         const sql = "INSERT INTO products(name, description, address, availability, category, user_id) VALUES($1, $2, $3, $4, $5, $6)";
-        const values = [name, description, address, availability, category, user_id];
+        const values = [name, description, address, availability, category, userId];
         return db.query(sql, values).then((dbRes) => dbRes.rows[0]);
     },
-    deleteProduct(id) {
-        const sql = "DELETE FROM products WHERE id = $1";
-        const values = [id];
+    deleteProduct(id, userId) {
+        const sql = "DELETE FROM products WHERE id = $1 AND user_id = $2";
+        const values = [id, userId];
+        return db.query(sql, values);
+    },
+    updateProductParameters(name, description, address, availability, product_id, userId) {
+        const sql = "UPDATE products SET name = $1, description = $2, address = $3, availability = $4 WHERE id = $6 AND user_id = $7";
+        const values = [name, description, address, availability, product_id, userId];
         return db.query(sql, values).then((dbRes) => dbRes.rows[0]);
     },
-    updateProductParameters(name, description, address, availability, product_id) {
-        const sql = "UPDATE products SET name = $1, description = $2, address = $3, availability = $4 WHERE id = $6";
-        const values = [name, description, address, availability, product_id];
-        return db.query(sql, values).then((dbRes) => dbRes.rows[0]);
-    },
-    getProductID(user_id) {
+    getProductID(userId) {
         const sql = "SELECT (id) FROM products WHERE user_id = $1;";
         return db
-            .query(sql, [user_id])
+            .query(sql, [userId])
             .then((dbRes) => dbRes.rows);
     },
     insertReview(review, rating, authorID, productID) {
